Add stayOnTop property to Label node
I added a boolean property to the Label node to set whether the Label should stay in front of other nodes by modifying LGraphCanvas.prototype.draw to call __LGraphCanvas.bringToFront after drawing nodes. 

I am very new to the Javascript side, I just wanted the labels to stay on top D: Changes are minimal and it works, but I was flying by the seat of my pants so please review for bad practices and/or a better method to do this.